Make base_analysis tables point to vr4mice.Dataset instead of vr4mice.VR4Mice
The base_analsysis schema isn't currently populated regardless.

This might just be temporary until the base_schemas intermediary schema is setup, but will stay like this for now so the module can be imported properly.

Also the Dataset and VR4Mice primary keys are the same, and VR4Mice didn't have any extra functionality that was used, besides that it linked to base_schemas.